Adding a very basic implementation of Store.
Idea is that we get the info from server in batches, but retrieve in pieces, and often from the very different points in Application. Sometimes it's just not worth to have a shared class for querying. It may be proxied through the Base class.

For instance, Builds were loaded through the Build Matrix, but they should be accessible from that moment from anywhere in UI.